+++ CHANGELOG.md
@@ -6,8 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [Unreleased]
 
-## [0.14.52] - 2021-01-05
+## [0.14.53] - 2021-01-05
 - Fixed `scheduled day only` option
+
+## [0.14.52] - 2021-01-07
+- Fix response submit
+
+## [0.14.51] - 2021-01-05
+- Fixed notification time
+- Added TokenLogger Prizes Activity feature
 
 ## [0.14.50] - 2021-01-05
 - Removed console logs